Use wl-copy to copy snippet to clipboard in `wl` mode
3c1ad70ec4b4865d2aafceb5847d783544d9c7ff introduced support for
wl-clipboard for better clipboard support on Wayland.

This implementation used `wl-paste` to **get** data from the clipboard,
but did use `wl-paste` to put data onto the clipboard, despite a
function for it being defined.

This commit changes the behavior to also use `wl-paste` to **put** data
on the clipboard.

This makes the implementation similar to xsel, which uses xsel for both
copy and paste when mode is set to xsel (instead of Gtk).