Add support for Markdown/HTML snippets
This adds support for snippets containing Markdown source which can be
rendered to HTML.

This allows applications that support rich-text/HTML clipboard contents
to insert HTML text, which is useful in WYSIWYG text editors, HTML email
clients, etc.

Markdown snippets are indicated by the presence of `markdown: true`
frontmatter and, optionally, a `markdown_extensions` field to tweak the
[Markdown Extensions] to enable.

Markdown snippets are only supported with the xsel/wl clipboard method,
as the `CopyToClipboardAction` does not support passing a mimetype.

[Markdown Extensions]: https://python-markdown.github.io/extensions/